paginator: Correct prev page offset
When paging, we may have a cursor that looks like such:

    <value=4, offset=0, is_prev=0>

And we have a dataset that looks like this:

    [ 1, 2, 3, 4, 5, 6, 7, 8]
               ^
      cursor value is here

When we ask for the prev-page cursor the prev-cursor will start from
value `4`, and do a lookup where the values are less than or equal to 4.
(The 'or equal' is used to generalize the query when doing next lookups)

The problem is when doing that 4 will be included, to fix this, prev
cursor should add a single offset to move pas the value that we had
already seen on the page.

The problem we're fixing here is that when *no value* is given in a
cursor, we move past a value that we haven't seen yet. This is fixed by
only applying the +1 offset for prev cursors that also have a value.